Remove legacy column-driven Workqueue components
Delete Workqueue, WorkqueueRowDesktop, WorkqueueRowMobile and
ListItemAction, which are fully superseded by the new WorkqueueRow.
The only remaining imports were enums and the NoResultText empty-state
component, which move to Workqueue/shared.ts so existing import paths
(@opencrvs/components/lib/Workqueue) keep working for Table,
UserAuditHistory and the v2-events client.
