+++ packages/components/src/Workqueue/index.ts
@@ -8,6 +8,6 @@
  *
  * Copyright (C) The OpenCRVS Authors located at https://github.com/opencrvs/opencrvs-core/blob/master/AUTHORS.
  */
-export * from './Workqueue'
+export * from './shared'
 export * from './types'
 export * from './components/WorkqueueRow'
